requirement.txt
django == 3.1
djangorestframework == 3.11.1
pillow == 7.2.0